pyproject.toml
[tool.poetry]
name = "registration"
version = "0.1.0"
description = ""
authors = ["Leonid Sliusar <loslyusar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.100.0"}
sqlalchemy = "^2.0.18"
alembic = "^1.11.1"
psycopg2-binary = "^2.9.6"
asyncpg = "^0.28.0"
bcrypt = "^4.0.1"
pydantic-settings = "^2.0.1"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
